base_de_dados/requirements.txt
PyMySQL==1.1.2
python-dotenv==1.2.1
types-PyMySQL==1.1.0.20250916
